Refactor modal detail view
Since there needs to be two types of detail views, it made more sense
to allow the data aspect to be handled by a super class and the visual
details to be handled by each of the subclasses, respectively.
