+++ Carleton150/TimelineDetailView.swift
@@ -2,10 +2,7 @@
 //  TimelineDetailView.swift
 //  Carleton150
 
-import Foundation
-
 class TimelineDetailView: UIViewController {
-    
     var parentView: UIViewController!
     var summary: String!
     var eventDescription: String?
@@ -13,9 +10,6 @@ class TimelineDetailView: UIViewController {
     var image: UIImage!
     var timestamp: String?
     
-    @IBOutlet weak var dateLabel: UILabel!
-    @IBOutlet weak var descriptionText: UITextView!
-   
     /**
         A convenience method for setting the data as a segue
         is initiated.
@@ -27,34 +21,5 @@ class TimelineDetailView: UIViewController {
         self.caption = timelineCell.cellCaption
         self.timestamp = timelineCell.cellTimestamp
         self.image = timelineCell.cellImage
-    }
-   
-    /**
-        Loads the view, darkens the background, and then sets the 
-        data inside the view.
-     */
-    override func viewDidLoad() {
-        // forces background to darken
-        self.view.backgroundColor = UIColor(white: 0, alpha: 0.6)
-       
-        // sets the current date and description
-        self.dateLabel.text = self.timestamp ?? ""
-        self.descriptionText.text = self.eventDescription
-        
-        // stops the text view from being edited
-        self.descriptionText.editable = false
-    }
-   
-    /**
-        Once the subviews are placed, the description text
-        needs to be set so that the text doesn't start at the bottom.
-     */
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-        self.descriptionText.setContentOffset(CGPointZero, animated: false)
-    }
-    
-    @IBAction func dismissDetailView(sender: AnyObject) {
-        parentView.dismissViewControllerAnimated(true) {}
     }
 }
